refactor(PAT-62): Apply SOLID - OPC  in Panels.
