+++ com/frontend/app/src/containers/ConsultationPanel.jsx
@@ -1,15 +1,8 @@
-/*********************************************************************************
-* In this component, we are using the Droppable component                          *
-* from react-beautiful-dnd to make the panel a drop zone for drag and drop cards.  *
-* We are also passing the droppableId and index properties to Droppable.           *
-* Also, we are using the CustomCard component to render each card within the panel.*
-***********************************************************************************/
 import React, { useEffect, useState } from 'react';
-import { Droppable } from 'react-beautiful-dnd';
 import Grid from '@mui/material/Grid';
-import { Paper } from '@mui/material';
 import TitlePanel from '../components/TitlePanel';
 import ConsultationTicket from './ConsultationTicket';
+import BasePanel from '../components/BasePanel';
 
 
 /**
@@ -22,21 +15,11 @@ import ConsultationTicket from './ConsultationTicket';
 const ConsultationPanel = ({ panel, index }) => {
   const [titleUpdateCount, setTitleUpdateCount] = useState(0); // force update view
   const [panelData, setPanelData] = useState(panel);
+  const title = TitlePanel({panel:panel});
 
   useEffect(() => {
     setPanelData(panel);
   }, [panel]);
-
-  const panelStyle = {
-    backgroundColor: panel.id === 0 ? '#87cefaab' : 'lightskyblue',
-    textAlign: 'center',
-    height: '75vh',
-    width: '20vw',
-  };
-
-  if (!panel) {
-    return <div>No panels.</div>;
-  }
 
   /**
    * Reduces the number of cards in the panel and updates the state.
@@ -47,22 +30,15 @@ const ConsultationPanel = ({ panel, index }) => {
     setTitleUpdateCount(titleUpdateCount + 1);
   };
 
+
   return (
-    <Droppable key={"droppeable-panel-"+String(panelData.id)} droppableId={String(panelData.id)} index={index} direction="vertical">
-      {(provided) => (
-        <Paper ref={provided.innerRef} {...provided.droppableProps}  style={panelStyle}>
-            <TitlePanel panel={panelData}/>
-            <Grid container  columns={12} spacing={2}  style={{width: '20vw', backgroundColor: '#d7f0fa' , flexDirection: 'column', margin: '0 auto'}} >
-                {panelData.cards.map((card, index) => (
-                    <Grid item xs={12} sm={6} md={11} key={card.consultation} >
-                    <ConsultationTicket card={card} index={index} reduce_number_cards={reduce_number_cards} key={card.consultation}/>
-                    </Grid>
-                ))}
-                {provided.placeholder}
-            </Grid>
-        </Paper>
-      )}
-    </Droppable>
+    <BasePanel panel={panel} index={index} title={title}>
+      {panelData.cards.map((card, index) => (
+          <Grid item xs={12} sm={6} md={11} key={card.consultation} >
+          <ConsultationTicket card={card} index={index} reduce_number_cards={reduce_number_cards} key={card.consultation}/>
+          </Grid>
+      ))}
+    </BasePanel>
   );
 };
